Update the foreach-statement key to use simple-variable
Previously the key of a foreach statement was specified as an
expression. In reality, php would not accept an arbitrary expression as
the key in a foreach loop. But it does appear to accept any
`simple-variable`, although it's difficult to come up with a valid use
case.

It might be mostly correct to accept only `variable-name` as the key,
but PHP will actually accept varvars and ${} expression variables as the
key.